UI: DTR screen: fix link: go to DOC view
