--- tests/end2end/screens/deep_traceability/view_document/go_from_requirement_card_to_requirement_in_document_view/input/document.sdoc
@@ -1,28 +1,0 @@
-[DOCUMENT]
-TITLE: Document title
-OPTIONS:
-  REQUIREMENT_STYLE: Table
-
-[FREETEXT]
-Hello world!
-[/FREETEXT]
-
-[SECTION]
-UID: SEC_UID
-TITLE: Section title
-
-[FREETEXT]
-Section text
-[/FREETEXT]
-
-[REQUIREMENT]
-UID: REC_UID
-TITLE: Requirement title
-STATEMENT: >>>
-Requirement statement.
-<<<
-RATIONALE: >>>
-Requirement rationale.
-<<<
-
-[/SECTION]
