+++ tests/end2end/screens/deep_traceability/view_document/view_document_go_from_requirement_card_to_requirement_in_document_view/expected_output/document2.sdoc
@@ -1,0 +1,12 @@
+[DOCUMENT]
+TITLE: Document #2 title
+
+[REQUIREMENT]
+UID: REQ-002
+TITLE: Requirement title
+STATEMENT: >>>
+Requirement statement.
+<<<
+RELATIONS:
+- TYPE: Parent
+  VALUE: REQ-001
