+++ tests/end2end/screens/deep_traceability/view_document/view_document_go_from_requirement_card_to_requirement_in_document_view/input/document.sdoc
@@ -1,0 +1,9 @@
+[DOCUMENT]
+TITLE: Document #1 title
+
+[REQUIREMENT]
+UID: REQ-001
+TITLE: Requirement title
+STATEMENT: >>>
+Requirement statement.
+<<<
